Use shuffle_group() directly
