+++ dask_cuda/explicit_comms/dataframe_merge.py
@@ -1,10 +1,10 @@
 import asyncio
 
 from dask.dataframe.core import _concat
+from dask.dataframe.shuffle import partitioning_index, shuffle_group
 from distributed.protocol import to_serialize
 
 from . import comms
-from .dataframe_shuffle import partition_by_hash
 
 
 async def send_df(ep, df):
@@ -65,6 +65,47 @@ async def broadcast(rank, root_rank, eps, df=None):
         return df
     else:
         return await recv_df(eps[root_rank])
+
+
+def partition_by_hash(df, columns, n_chunks, ignore_index=False):
+    """Splits dataframe into partitions
+
+    The partitions is determined by the hash value of the rows in `columns`.
+
+    Parameters
+    ----------
+    df: DataFrame
+    columns: label or list
+        Column names on which to split the dataframe
+    npartition: int
+        Number of partitions
+    ignore_index : bool, default False
+        Set True to ignore the index of `df`
+
+    Returns
+    -------
+    out: Dict[int, DataFrame]
+        A dictionary mapping integers in {0..npartition} to dataframes.
+    """
+    if df is None:
+        return [None] * n_chunks
+
+    # Hashing `columns` in `df` and assing it to the "_partitions" column
+    df["_partitions"] = partitioning_index(df[columns], n_chunks)
+    # Split `df` based on the hash values in the "_partitions" column
+    try:
+        # For Dask < 2.17 compatibility
+        ret = shuffle_group(df, "_partitions", 0, n_chunks, n_chunks, ignore_index)
+    except TypeError:
+        ret = shuffle_group(
+            df, "_partitions", 0, n_chunks, n_chunks, ignore_index, n_chunks
+        )
+
+    # Let's remove the partition column and return the partitions
+    del df["_partitions"]
+    for df in ret.values():
+        del df["_partitions"]
+    return ret
 
 
 async def hash_join(n_chunks, rank, eps, left_table, right_table, left_on, right_on):
